+++ up42/viztools.py
@@ -1,7 +1,0 @@
-"""
-Visualization tools available in various objects
-"""
-
-
-class VizTools:
-    pass